requirements.txt
bcrypt==3.2.0
email-validator==1.1.3
fastapi==0.73.0
passlib==1.7.4
pydantic==1.9.0
pymongo==4.0.1
python-jose==3.3.0
python-multipart==0.0.5
sha512-crypt==0.1.1
sniffio==1.2.0
typing_extensions==4.0.1
uvicorn==0.17.1
python-dotenv==0.21.0